CS-744Assign1: Implement sorting of IOT device data on Spark
[Problem]
Need to sort the IOT data based on country code first.
If country codes are same, then we need to sort based on timestamp.

[Solution]
Loaded the input CSV file onto HFDS.
Loaded the input data into a dataframe and sorted the data
according to the above condition.
Wrote the sorted output into the specified CSV folder.
Solutions are available in both Scala and Python.

[Test]
Tested that the sorted output is visible in the output folder

[TODO]
1. Write run.sh for environment setup.
2. Cleanup code and add comments.
3. If needed, need to write the output into single CSV
file rather than to a folder.
4. Currently, the scala code is not yet fully configured
to handle dynamic input. Need to write a wrapper script
on top of spark-submit to handle this.